Externalize DB generic utility functions
